feat: modify delete cost_reports_data location
